requirements.txt
asgiref==3.8.1
cfgv==3.4.0
commonmark==0.9.1
distlib==0.3.8
dj-database-url==2.2.0
Django==5.1.1
djangorestframework==3.15.2
Faker==30.6.0
filelock==3.16.1
gunicorn==23.0.0
identify==2.6.1
nodeenv==1.9.1
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pre-commit==3.8.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
virtualenv==20.26.6
whitenoise==6.7.0
